Refactor DiseaseMapper, DiseaseRepo, and DiseaseService to use Long for diseaseId; update mapping for DiseaseResponse and add new service methods

+++ src/main/java/com/schoolhealth/schoolmedical/repository/DiseaseRepo.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface DiseaseRepo extends JpaRepository<Disease, Integer> {
+public interface DiseaseRepo extends JpaRepository<Disease, Long> {
     @Query(value = "SELECT * FROM disease d where d.is_active =?1", countQuery = "SELECT COUNT(*) FROM disease d where d.is_active =?1", nativeQuery = true)
     Page<Disease> findAllByisActiveTrue(boolean isActive, Pageable pageable);
